% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_emu.R
\name{make_emu}
\alias{make_emu}
\title{make_emu: build emulator}
\usage{
make_emu(designX, responseF, forcingX, r = NULL, thresh = 0.999)
}
\arguments{
\item{designX}{Full dataset design}

\item{responseF}{Full dataset response}

\item{forcingF}{Full climate timeseries for SVD (optional)}
}
\value{
`make_emu()` returns an emulator object to use.
}
\description{
Build emulators of principal components with RobustGaSP.
Based on code by Jonty Rougier.
}
