% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVDimpute.R
\name{SVDimpute}
\alias{SVDimpute}
\title{SVD Imputation Algorithm for Matrix Completion}
\usage{
SVDimpute(X, k = NULL, pmin = 1 - 1e-04, maxit = 5)
}
\arguments{
\item{X}{numeric matrix containing missing values as `NA`.}

\item{k}{rank of approximation to `X`.}

\item{pmin}{minimum proportional variation, used to set `k` if `k = NULL`, see Details.}

\item{maxit}{maximum number of iterations.}
}
\value{
A matrix like `X` but with the `NA`s imputed, plus an attribute `"param"`, a list with `pmin`, `k`, and `rank` if `maxit > 0`.
}
\description{
Replace `NA` values in matrix `X` with imputed values.  Imputation is along the rows, assuming that all columns of `X` have the same units.

_The order of the columns matters!_ The initial infill is by a variant on Last Value Carried Forward (LVCF).  Therefore it is important that the columns follow a sequence.  If the columns represent points in a grid, then use [snake()] to reorder them before imputation.
}
\details{
This is a two-stage imputation.  Stage 1 is an initial infill of all `NA`s using a variant on LVCF.  If `X[i,j]` is `NA`, it is filled by `X[i,j-1] + diff(X[istar, c(j-1,j)])` where `istar` is the run with the closest value to `X[i,j-1]`.This is followed by a similar backward pass, and the function terminates with an error if there are `NA` remaining after both passes.

Stage 2 (if `maxit > 0`) is a SVD imputation, similar to that described in Troyanskaya et al (2001), see References.  This is carried out after centering `X` (now with no `NA`s).  If `k` is not specified, it is the rank of the SVD approximation of `X` which achieves a proportion `pmin` of the variation of `X`.
}
\examples{
## demo using diamondEOO

Y <- diamondEOO$Y
mask <- is.na(Y)
show(mean(mask)) # 6\% NAs where P < 0, at high values of V

## LVCF only

Y1 <- SVDimpute(Y, maxit = 0)
show(summary(Y1[mask])) # all small
all.equal(Y[!mask], Y1[!mask]) # TRUE

## now with some SVD iterations

Y1 <- SVDimpute(Y, maxit = 5) # the default
show(summary(Y1[mask])) # all small

## using snake()

dim <- lengths(diamondEOO$grid)
oo <- snake(dim)
Y2oo <- SVDimpute(Y[, oo])
Y2 <- Y2oo[, order(oo)]
show(summary((Y1 - Y2)[mask])) # minor differences
}
\references{
Olga Troyanskaya, Michael Cantor, Gavin Sherlock, Pat Brown, Trevor Hastie, Robert Tibshirani, David Botstein, Russ B. Altman, Missing value estimation methods for DNA microarrays, Bioinformatics, Volume 17, Issue 6, June 2001, Pages 520–525, <https://doi.org/10.1093/bioinformatics/17.6.520>.
}
\seealso{
LVCF will give a poor initial fill for points on the upper edge of a grid.  If the columns of `X` represent points in a grid, then reorder them using the permutation from [snake()] before imputation.
}
